Refactor File
The location ID request should have been happening in a different method.

The location ID request now will raise an HTTP error when it fails, that might be partially handled by the server in the future.

In general the code should be much cleaner and easier to read without raising a false exeception.